third commit

+++ backend/database/migrations/2025_02_20_082830_create_blogs_table.php
@@ -11,14 +11,16 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('blogs', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('user_id')->constrained('users')->onDelete('cascade');
-            $table->string('authorName')->nullable();
-            $table->string('title')->nullable();
-            $table->text('content')->nullable();
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('blogs')) {
+            Schema::create('blogs', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('user_id')->constrained('users')->onDelete('cascade');
+                $table->string('authorName')->nullable();
+                $table->string('title')->nullable();
+                $table->text('content')->nullable();
+                $table->timestamps();
+            });
+        }
     }
 
     /**
